+++ src/signer/canister/tests/it/eth.rs
@@ -52,7 +52,7 @@ mod sign_transaction {
             owner: test_env.signer.canister_id(),
             subaccount: None,
         };
-        let amount: u64 = SignerMethods::EthSignTransaction.fee() + LEDGER_FEE as u64;
+        let amount: u128 = SignerMethods::EthSignTransaction.fee() + LEDGER_FEE;
         test_env
             .ledger
             .icrc_2_approve(caller, &ApproveArgs::new(payment_recipient, amount.into()))
@@ -123,7 +123,7 @@ mod personal_sign {
             owner: test_env.signer.canister_id(),
             subaccount: None,
         };
-        let amount: u64 = SignerMethods::EthPersonalSign.fee() + LEDGER_FEE as u64;
+        let amount: u128 = SignerMethods::EthPersonalSign.fee() + LEDGER_FEE;
         test_env
             .ledger
             .icrc_2_approve(caller, &ApproveArgs::new(payment_recipient, amount.into()))
@@ -198,7 +198,7 @@ mod eth_address {
             owner: test_env.signer.canister_id(),
             subaccount: None,
         };
-        let amount: u64 = SignerMethods::EthAddress.fee() + LEDGER_FEE as u64;
+        let amount: u128 = SignerMethods::EthAddress.fee() + LEDGER_FEE;
         test_env
             .ledger
             .icrc_2_approve(caller, &ApproveArgs::new(payment_recipient, amount.into()))
